+++ doc/changelog.rst
@@ -68,6 +68,7 @@ Localization
 Admin
 ~~~~~
 
+- Add target option to ``SearchResult``
 - Filter contacts with group in contact list view
 - Hide group members from contact group edit view
 - Remove support for select pickers
@@ -172,6 +173,12 @@ Customer Group Pricing
 
 - Rename Simple Pricing to Customer Group Pricing
 - Fix pricing for ``AnonymousContact``
+
+
+Guide
+~~~~~
+
+- Add guide app that integrates help documentation into admin search
 
 Discount Pricing
 ~~~~~~~~~~~~~~~~